functions, cleaner+manual nginx and ssl config

+++ add-website.sh
@@ -1,29 +1,88 @@
 #!/bin/bash
 
+function restart_nginx() {
+    nginx -t > /dev/null
+    if [ $? -ne 0 ]; then
+        echo "Problem with generating nginx config. Quitting."
+        nginx -t
+        exit 1
+    fi
+    nginx -s reload
+}
+
+function enable_ssl() {
+    domain="$1"
+    config_file="$2"
+
+    # Check DNS
+    dns=$(host "$domain")
+    if [ $? -ne 0 ]; then
+        echo "DNS record for $domain does not exist"
+        exit 1
+    fi
+
+    # Optain certificate
+    certbot certonly --nginx -d "$domain"
+
+    # Replace nginx config with SSL version
+    if [ $? -eq 0 ]; then
+        # Enable https config
+        sed -i 's/listen 80;/listen 443 ssl http2;/g' "$config_file"
+        sed -i 's/listen \[::\]:80;/listen [::]:443 ssl http2;/g' "$config_file"
+        sed -i 's/# include/include/g' "$config_file"
+        sed -i 's/# ssl_certificate/ssl_certificate/g' "$config_file"
+        sed -i 's/# ssl_certificate_key/ssl_certificate_key/g' "$config_file"
+
+        # Append HTTP -> HTTPS redirection
+cat >> "$config_file" <<EOF
+
+# Redirect HTTP to HTTPS
+server {
+    listen 80;
+    listen [::]:80;
+
+    server_name $domain;
+
+    if (\$host = $domain) {
+        return 301 https://\$host\$request_uri;
+    }
+
+    return 404;
+}
+
+EOF
+fi
+
+restart_nginx
+}
+
 read -r -p "Domain: " domain
 config_file="/etc/nginx/conf.d/$domain.conf"
 
 
 # If config file exists, ask to remove site
 if [ -f "$config_file" ]; then
-    read -r -p "$config_file already exists. What would you like to do? [r(emove) / c(ertificate) / Q(uit)]: " remove
+    if [[ -d "/etc/letsencrypt/live/$domain" ]]; then
+        read -r -p "$config_file already exists. What would you like to do? [r(emove) / Q(uit)]: " remove
+    else
+        read -r -p "$config_file already exists. What would you like to do? [r(emove) / c(ertificate) / Q(uit)]: " remove
+    fi
+
     if [[ $remove == "r" || $remove == "r" ]]; then
         dns=$(host "$domain")
-        if [ $? == 0 ]; then
+        if [[ $? == 0 && -d "/etc/letsencrypt/live/$domain" ]]; then
             certbot revoke --delete-after-revoke --cert-name "$domain"
         fi
-        rm -r "/var/www/$domain"
-        rm -r "/etc/nginx/conf.d/$domain.conf"
-        systemctl restart nginx
-    elif [[ $remove == "c" || $remove == "C" ]]; then
-        # Check DNS
-        dns=$(host "$domain")
-        if [ $? -ne 0 ]; then
-            echo "DNS record for $domain don't exist"
-            exit 1
+
+        if [ -d "/var/www/$domain" ]; then
+            rm -r "/var/www/$domain"
         fi
-        # Optain certificate
-        certbot --redirect --nginx -d "$domain"
+
+        rm -r "/etc/nginx/conf.d/$domain.conf"
+
+        restart_nginx
+    elif [[ $remove == "c" || $remove == "C" ]]; then
+        enable_ssl "$domain" "$config_file"
     else
         echo "Quitting"
     fi
@@ -42,18 +101,25 @@ read -r -p "Get Let's Encrypt certificate? [y/N]: " get_cert
 # Create nginc config file
 cat > "$config_file" <<EOF
 server {
-    listen 80;
     listen [::]:80;
+    listen 80;
+
     server_name $domain;
 
-    error_log /var/log/nginx/$domain.error.log debug;
-    access_log /var/log/nginx/$domain.access.log;
-    
     location / {
         proxy_pass http://localhost:$port;
         proxy_set_header X-Forwarded-For ]\$remote_addr;
     }
+
+    # include /etc/nginx/options.conf;
+
+    access_log /var/log/nginx/$domain.access.log;
+    error_log /var/log/nginx/$domain.error.log;
+
+    # ssl_certificate /etc/letsencrypt/live/$domain/fullchain.pem;
+    # ssl_certificate_key /etc/letsencrypt/live/$domain/privkey.pem;
 }
+
 EOF
 
 
@@ -73,16 +139,13 @@ read -r -p "Get Let's Encrypt certificate? [y/N]: " get_cert
 # Create nginx config file
 cat > "$config_file" <<EOF
 server {
-    listen 80;
     listen [::]:80;
+    listen 80;
 
     server_name $domain;
 
     root /var/www/$domain/html;
     index $indexes;
-
-    error_log /var/log/nginx/$domain.error.log debug;
-    access_log /var/log/nginx/$domain.access.log;
 
 EOF
 
@@ -117,27 +180,27 @@ then
 
 cat >> "$config_file" <<EOF
 
-    # Enable CORS headers
-    if (\$request_method = 'OPTIONS') {
-        add_header 'Access-Control-Allow-Origin' '*';
-        add_header 'Access-Control-Allow-Methods' 'GET, POST, OPTIONS';
-        add_header 'Access-Control-Max-Age' 1728000;
-        add_header 'Content-Type' 'text/plain; charset=utf-8';
-        add_header 'Content-Length' 0;
-        return 204;
-    }
-    if (\$request_method = 'POST') {
-        add_header 'Access-Control-Allow-Origin' '*';
-        add_header 'Access-Control-Allow-Methods' 'GET, POST, OPTIONS';
-        add_header 'Access-Control-Allow-Headers' 'DNT,User-Agent,X-Requested-With,If-Modified-Since,Cache-Control,Content-Type,Range';
-        add_header 'Access-Control-Expose-Headers' 'Content-Length,Content-Range';
-    }
-    if (\$request_method = 'GET') {
-        add_header 'Access-Control-Allow-Origin' '*';
-        add_header 'Access-Control-Allow-Methods' 'GET, POST, OPTIONS';
-        add_header 'Access-Control-Allow-Headers' 'DNT,User-Agent,X-Requested-With,If-Modified-Since,Cache-Control,Content-Type,Range';
-        add_header 'Access-Control-Expose-Headers' 'Content-Length,Content-Range';
-    }
+        # Enable CORS headers
+        if (\$request_method = 'OPTIONS') {
+            add_header 'Access-Control-Allow-Origin' '*';
+            add_header 'Access-Control-Allow-Methods' 'GET, POST, OPTIONS';
+            add_header 'Access-Control-Max-Age' 1728000;
+            add_header 'Content-Type' 'text/plain; charset=utf-8';
+            add_header 'Content-Length' 0;
+            return 204;
+        }
+        if (\$request_method = 'POST') {
+            add_header 'Access-Control-Allow-Origin' '*';
+            add_header 'Access-Control-Allow-Methods' 'GET, POST, OPTIONS';
+            add_header 'Access-Control-Allow-Headers' 'DNT,User-Agent,X-Requested-With,If-Modified-Since,Cache-Control,Content-Type,Range';
+            add_header 'Access-Control-Expose-Headers' 'Content-Length,Content-Range';
+        }
+        if (\$request_method = 'GET') {
+            add_header 'Access-Control-Allow-Origin' '*';
+            add_header 'Access-Control-Allow-Methods' 'GET, POST, OPTIONS';
+            add_header 'Access-Control-Allow-Headers' 'DNT,User-Agent,X-Requested-With,If-Modified-Since,Cache-Control,Content-Type,Range';
+            add_header 'Access-Control-Expose-Headers' 'Content-Length,Content-Range';
+        }
 EOF
 
 fi # Ending if $cors
@@ -165,9 +228,18 @@ EOF
 # fi # Ending if $use_php
 
 
-# Ending block server {
-echo "}" >> "$config_file"
+# Ending block "server {"
+cat >> "$config_file" << EOF
 
+    # include /etc/nginx/options.conf;
+
+    access_log /var/log/nginx/$domain.access.log;
+    error_log /var/log/nginx/$domain.error.log;
+
+    # ssl_certificate /etc/letsencrypt/live/$domain/fullchain.pem;
+    # ssl_certificate_key /etc/letsencrypt/live/$domain/privkey.pem;
+}
+EOF
 
 # Create /var/www stuff
 mkdir -p "/var/www/$domain/html"
@@ -200,28 +272,10 @@ cat > "/var/www/$domain/html/index.html" <<EOF
 </html>
 EOF
 
-fi # Ending the else on if $proxy
+fi # Ending the "else" on "if $proxy"
 
+restart_nginx
 
-# Test nginx config
-nginx -t > /dev/null
-if [ $? -ne 0 ]; then
-    echo "Problem with generating nginx config. Quitting."
-    nginx -t
-    exit 1
-fi
-systemctl restart nginx
-
-
-# Check DNS
-dns=$(host "$domain")
-if [ $? -ne 0 ]; then
-    echo "DNS record for $domain don't exist"
-    exit 1
-fi
-
-
-# Get and enable LetsEncrypt SSL cert
-if [[ "$get_cert" == "y" || "$get_cert" == "Y" ]]; then
-    certbot --redirect --nginx -d "$domain"
+if [[ $get_cert == "y" || $get_cert == "Y" ]]; then
+    enable_ssl "$domain" "$config_file"
 fi
